Added local fetch call to json file for Pokemon - Added styling rules and new section for name and type
This adds a few changes, mainly the javascript file. This will be an onload of the document. Will fetch the local json document and then grab a few details from it then display it in the appropriate places. have added a few styling rules for the image just so it doesn't go out of proportion.